fixed that not all data were received from conUpdateInfo()
git-svn-id: https://subversion.assembla.com/svn/atarijookie/trunk@712 0d2ca917-fb58-4453-bb26-a698c01b0940

+++ cosmosex/network/netadapter.h
@@ -66,7 +66,7 @@ public:
     }
 
     BYTE *data;
-    BYTE count;
+    WORD count;
 };
 
 //-------------------------------------
